requirements.txt
flask
requests
python-dotenv
python-binance
pandas
numpy
scikit-learn
openai